Update quiz.html
I'm very new to Java so I do apologize if I am incorrect here. But I believe the issue with the quiz ending after five questions is that the if/else query is deciding if you are on question number five and, if so, ending the quiz. This should be deciding if you are on question number 70.